% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{confint.myLm}
\alias{confint.myLm}
\title{STAT 6210 Least squares regression estimator}
\usage{
\method{confint}{myLm}(x, alpha = 0.05, approach = "asymp")
}
\arguments{
\item{x}{\code{x} Output data from myLm.}

\item{alpha}{\code{alpha} Specified alpha level}

\item{approach}{\code{approach} Can specify Asym or Boot}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{lwr}{lower confidence bound}
     \item{upr}{upper confidence bound}
}
}
\description{
Calculates confidence intervals for estimated coefficients from
regression analysis in myLm function.
}
\examples{
library(MASS)
data(Boston)
fit = myLm(Boston$crim,Boston[c("age", "medv")])
# Compute asymptotic confidence interval with alpha of 0.05
confint(fit)

}
\author{
Todd Steury, Luke Dolan, and Kerry Cobb
}
