% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{myLm}
\alias{myLm}
\title{STAT 6210 Least squares regression estimator}
\usage{
myLm(response, covariates)
}
\arguments{
\item{response}{\code{vector} of dimension 1 that represents the y values
for the regression analysis.}

\item{covariates}{\code{matrix} vector, matrix, or data.frame representing
the x variables for the regression analysis}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{betas}{Coefficient estimates}
     \item{sigma2}{Residual variance}
     \item{variance_beta}{Variance in coefficient estimates}
     \item{residuals}{Vector of residuals}
     \item{predictors}{Matrix of x data}
     \item{y}{Vector of response values}
     \item{df}{degrees of freedom used in calculations}
     \item{yHat}{Predicted y values from regression equation}
     \item{MSPE}{Mean squared prediction error}
     \item{SSM}{Sum of Squares from regression}
     \item{SSE}{Sum of squared error}
     \item{FStatistic}{F statistic for ANOVA test}
     \item{Pvalue}{P-value from F test}
}
}
\description{
Uses least squares matrix calculations to estimate coefficients
from regression analysis.
}
\examples{
library(MASS)
data(Boston)
fit = myLm(Boston$crim,Boston[c("age", "medv")])
fit
}
\author{
Todd Steury, Luke Dolan, and Kerry Cobb
}
